requirements/requirements.txt
asgiref==3.2.6
Django==3.0.4
django-ckeditor==5.9.0
django-js-asset==1.2.2
jdatetime==3.6.2
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
